requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
yfinance==0.2.18
fredapi==0.5.1
python-dotenv==1.0.0
